Broke tests out into individual files and updated README
