+++ src/main/java/liuni/KeyHandler.java
@@ -55,7 +55,7 @@ public class KeyHandler {
         catch (Exception e) {
             System.out.println("Could not set up the keys. " + e.toString());
             e.printStackTrace();
-//            System.exit(-2);
+            System.exit(-2);
         }
     }
 
@@ -78,7 +78,7 @@ public class KeyHandler {
         catch (IOException e) {
             System.out.println("Error occurred when setting up twitter4j.properties.");
             e.printStackTrace();
-//            System.exit(-2);
+            System.exit(-2);
         }
         finally {
             try {
@@ -90,7 +90,7 @@ public class KeyHandler {
             catch (IOException e) {
                 System.out.println("Error occurred when closing the Buffered writer.");
                 System.out.println(e.toString());
-//                System.exit(-2);
+                System.exit(-2);
             }
         }
     }
