+++ src/test/java/liuni/ResponseListImpl.java
@@ -1,0 +1,22 @@
+package liuni;
+
+import twitter4j.RateLimitStatus;
+import twitter4j.ResponseList;
+
+import java.util.ArrayList;
+
+class ResponseListImpl<T> extends ArrayList<T> implements ResponseList<T> {
+    private static final long serialVersionUID = 9105950888010803544L;
+    private transient RateLimitStatus rateLimitStatus = null;
+    private transient int accessLevel;
+
+    @Override
+    public RateLimitStatus getRateLimitStatus() {
+        return rateLimitStatus;
+    }
+
+    @Override
+    public int getAccessLevel() {
+        return accessLevel;
+    }
+}
